feat: implement static CONFIG system with zero-cost access
- Replace dynamic config loading with LazyLock static CONFIG pattern
- Add serde "rc" feature for Arc<Vec<T>> serialization support
- Unify config structures in config::core module
- Integrate sync module to work directly with CONFIG.sync.repos
- Simplify CLI config commands to show-only functionality
- Update hook executor to use static CONFIG internally
- Remove sync_backup module and unnecessary config complexity
- Add Clone derive to all Args structs for efficient CLI handling
- Fix all compilation errors and type mismatches

Performance improvements:
- Zero-cost config access via static reference
- Arc-based sharing eliminates vector cloning
- LazyLock ensures single initialization

Breaking changes:
- Removed config get/set/init/validate CLI commands
- HookExecutor no longer takes config parameter
- Simplified sync configuration structure

+++ packages/guardy/src/parallel/core.rs
@@ -317,7 +317,7 @@ impl ExecutionStrategy {
     /// - Application logic
     ///
     /// Domain-specific adaptations should be handled by the calling module.
-    pub fn calculate_optimal_workers(max_threads_config: usize, thread_percentage: u8) -> usize {
+    pub fn calculate_optimal_workers(max_threads_config: u16, thread_percentage: u8) -> usize {
         let available_cores = num_cpus::get();
 
         // Calculate workers based on percentage of available cores
@@ -326,7 +326,7 @@ impl ExecutionStrategy {
 
         // Apply config limit if specified (0 means use percentage calculation only)
         if max_threads_config > 0 {
-            std::cmp::min(max_threads_config, workers_by_percentage)
+            std::cmp::min(max_threads_config as usize, workers_by_percentage)
         } else {
             workers_by_percentage
         }
